fix swarm

+++ auto-swarm.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
-source /etc/hlm-cfgs/icnp.cfg
-if [ "$nodeurl" != "myhornetnode.ddns.net" ]; then
-    curl --silent --output /dev/null -X DELETE https://register.tanglebay.org/nodes/$nodepassword
-    curl -X POST "https://register.tanglebay.org/nodes" -H  "accept: */*" -H  "Content-Type: application/json" -d "{ \"name\": \"$nodename\", \"url\": \"https://$nodeurl:$nodeport\", \"address\": \"$donationaddress\", \"pow\": \"$pownode\", \"password\": \"$nodepassword\" }" |jq
+source /etc/hlm-cfgs/swarm.cfg
+if [ -f "/etc/hlm-cfgs/swarm.cfg" ]; then
+    if [ "$nodeurl" != "myhornetnode.ddns.net" ]; then
+        curl --silent --output /dev/null -X DELETE https://register.tanglebay.org/nodes/$nodepassword
+        curl -X POST "https://register.tanglebay.org/nodes" -H  "accept: */*" -H  "Content-Type: application/json" -d "{ \"name\": \"$nodename\", \"url\": \"https://$nodeurl:$nodeport\", \"address\": \"$donationaddress\", \"pow\": \"$pownode\", \"password\": \"$nodepassword\" }" |jq
+    fi
 fi
 exit 0
